fix Bottom Navigation
