+++ app/src/main/java/com/example/ncast/ui/home/HomeFragment.kt
@@ -1,0 +1,20 @@
+package com.example.ncast.ui.home
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.ncast.R
+
+class HomeFragment : Fragment(R.layout.fragment_home) {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+}
