+++ app/src/main/java/com/example/ncast/ui/profile/ProfileFragment.kt
@@ -1,0 +1,19 @@
+package com.example.ncast.ui.profile
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.ncast.R
+
+class ProfileFragment : Fragment(R.layout.fragment_profile) {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_profile, container, false)
+    }
+}
