+++ app/src/main/res/menu/bottom_menu.xml
@@ -1,26 +1,21 @@
-<?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-
-
     <item
-        android:id="@+id/bottom_home"
+        android:id="@+id/homeFragment"
         android:icon="@drawable/baseline_home_24"
         android:title="Home" />
 
     <item
-        android:id="@+id/bottom_search"
+        android:id="@+id/searchFragment"
         android:icon="@drawable/baseline_search_24"
         android:title="Search" />
 
     <item
-        android:id="@+id/bottom_settings"
+        android:id="@+id/libraryFragment"
         android:icon="@drawable/baseline_library_music_24"
         android:title="Library" />
 
     <item
-        android:id="@+id/bottom_profile"
+        android:id="@+id/profileFragment"
         android:icon="@drawable/baseline_person_24"
         android:title="Profile" />
-
-
 </menu>
